Adicao de dicas em pontos turisticos.

+++ viajantes/app/helpers/application_helper.rb
@@ -209,6 +209,9 @@ module ApplicationHelper
 		case controller_name.intern
 			when :shops
 				return shop_shop_evaluation_path(params[:id],evaluation.especified_type)
+				
+			when :tourist_sights
+				return tourist_sight_tourist_sight_evaluation_path(params[:id],evaluation.especified_type)
 		end
 	end
 	
@@ -216,6 +219,11 @@ module ApplicationHelper
 		case controller_name.intern
 			when :shops
 				return {:action => "destroy", :controller => :shop_evaluations,
+							  :confirm => 'Você tem certeza?', :method => :delete,
+							  :title => "Apagar"}
+							  
+			when :tourist_sights
+				return {:action => "destroy", :controller => :tourist_sights_evaluations,
 							  :confirm => 'Você tem certeza?', :method => :delete,
 							  :title => "Apagar"}
 		end
